Fix init script for debian Buster
start-stop-daemon since version 1.19.3 fail to stop
matching only a pidfile that can be writen by an unprivileged
(non-root) user because it is considered a security risk.

Adding a second matching option (user) solved this.
This fix does not break compatibility with older versions
and does not mess with pidfile's mode or permissions.

Signed-off-by: Vicente Jimenez Aguilar <googuy@gmail.com>